requirements.txt
asgiref==3.3.0
Authlib==0.15.3
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
colour==0.1.5
cryptography==3.4.2
Django==3.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
djangorestframework==3.12.2
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==3.1
importlib-metadata==2.0.0
numpy==1.19.4
pandas==1.1.4
pycparser==2.20
pyreadline==2.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
sqlparse==0.4.1
tda-api==1.1.4
urllib3==1.26.3
websockets==8.1
zipp==3.4.0
